Improve validation of rule options in schema files
Error messages have been added to return statements in the AllMustContain, AtLeastContains, StrEndsWith, and StrStartsWith rules when no value or an empty value is provided in the schema files. This results in more informative feedback. Additionally, corresponding tests have been instituted in the RulesTest.php file to verify the functioning of these updates.