New validation rules and refactoring (#23)
This commit updates the GitHub Actions workflows by adding a new
"test-phar" job for verifying the functionality of the PHAR binary. The
names and actions of existing jobs were updated to be more informative.
The README file was also cleaned up for better readability. Changes also
included the addition of planned features related to tool versioning.