terminal width auto-detection optimization
Removed the assignment of the terminal width auto-detection to a variable in ErrorSuite.php and instead directly used it in the limit calculation. To improve clarity and maintain cleanliness of the code, comments related to terminal width auto-detection fall back have been moved to inside the autoDetectTerminalWidth function in Utils.php.