Refactor validate method and remove unused import statement
The validate method in AbstractRule.php file is now able to accept either an array or string as the cell value, improving flexibility in validation. Unused import statement for Utils class has also been removed to increase efficiency in the code. Additionally, a runtime exception has been incorporated to handle the scenario when the 'validateRule' method is not found, thus, increasing the robustness of the code.