Implement fallback line number in validators
Introduced constant FALLBACK_LINE as a fallback line number to be used in validators to avoid using magic number -1. With this modification, all line validation error references will now be to a valid line number in the source files, which improves traceability in reports. Furthermore, test descriptions were also adjusted to reflect this new behavior.