Aggregate rules and refactor rules structure (#27)
The rules structure for the Csv-Blueprint project is reorganized in this
change, to better differentiate between "Cell Rules" and "Aggregate
Rules". With this change, all former rules are moved into a dedicated
rules folder, highlighting their specific focus. This structural change
improves the readability of the project and facilitates future additions
to the cell rules or aggregate rules.