Refine floating-point comparison in aggregation rules
Improved the precision and consistency of floating-point comparisons, especially in aggregation rules. Changes include suppressing PhanUnusedProtectedMethodParameter in AbstractAggregateRule.php and adjustments to AbstractAggregateRuleCombo.php. Now, before a strict comparison, the numbers are rounded to 10 decimal places because of the inherent imprecision of floating-point arithmetic in computers. This enhances the reliability of equality checks. Adjusted test cases