Refactor average calculation to use MathPHP library
The average calculation in the ComboAverage rule has been refactored to use the mean function from the MathPHP library instead of manually calculating it. This update helps to improve the accuracy of the calculation and handle any exceptions by returning 0. The installation of "markrogoyski/math-php" has been adjusted in the composer.json file.