Add `count` aggregate rule
Introduced a new aggregate rule, ComboCount in Csv-Blueprint. This counts the total number of rows in the CSV file. Also updated the existing 'average' aggregate rules to convert array values to float before calculation. The schema examples and README are updated to demonstrate the new 'count' aggregate rule. A new test case for the 'count' rule has also been added.