Update handling of null aggregate values (#43)
Updated methods to handle null values for aggregate calculations by
changing the return type of the getActualAggregate function to '?float',
so that it can now allow null values. This change provides more
flexibility by not forcing a 0 value return when the aggregate
calculation is not possible and aids in better error management and
debugging.