Added `not_allow_values` rule
Introduced a new NotAllowValues rule that indicates the set of values which are not allowed. This rule was added to full.json, full.php, and full.yml files. Accommodating this new rule required the creation of corresponding classes and test files. The enhancement provides a more robust and configurable data validation effort.