Add `is_json` rule (#50)
The commit introduces a `is_json` rule to validate if a given input is a
valid JSON. This rule has been added to the json, php, and yml schema
examples. Accompanying unit tests have also been added. The rule count
in the README.md file is also updated to reflect the new addition.