Add `population_variance` to aggregate rules
Introduced the "Population Variance" rule to the set of aggregate rules, updating examples and tests. Implemented in the schema examples for json, php, and yml files. Also created a new PHP class for the rule and its corresponding unit tests, while updating the README to reflect the total count of implemented rules.