Change yml to phpArray in Schema validation
During the schema validation process in src/Schema.php, the function used to get the expected validity of CSV files was updated. Instead of using 'yml()', 'phpArray()' method is now used that matches the file type of 'full.php'. This change improves the alignment with the used schema files and the validation process.