Improve Docker setup and update dependencies
In this update, a custom php.ini file is introduced to configure the PHP settings in Docker. The Dockerfile has been altered to include this file. The dependencies of JBZoo/Utils and others in composer.lock and composer.json were updated. Moreover, changes were made to the Makefile to better illustrate the executables and demo steps.