Simplify GA workflow (#75)
This commit includes updates to GitHub workflows and dependencies. The
adjustments in workflows corresponds to alteration in triggering
conditions and updating version of 'actions/checkout' while in the
dependencies, corrections have been made to change versions for
different libraries and tools. Some of the changes seem to downgrade the
version which could be to resolve compatibility issues.