Add "contains_none" rule to cell validation (#88)
This update introduces a new rule named "contains_none" to the cell
validation process. This rule takes an array of strings and validates
that none of these strings appear in the cell value. Changes have been
incorporated in the relevant schema example files and the associated PHP
and Test files for the new validation rule have been created.