Add 'is_sorted' rule to aggregate rules
The 'is_sorted' rule has been added to the aggregate rules section of the schema examples in various formats (json, php, yml, etc.). This rule checks if the column is sorted in a specific order. An implementation of the rule has also been added to the 'Rules/Aggregate' section of the application's source code and tests for this rule have been written under 'tests/Rules/Aggregate/SortedTest.php'.