Rename SortedTest to IsSortedTest
The SortedTest class and its file have been renamed to IsSortedTest to better reflect its functionality. This class tests whether a column is sorted according to the 'is_sorted' rule in the aggregate rules section of the codebase. This change aligns the test class name with the name of the rule it verifies.