Add version for output
Enhanced the software to include a versioning feature that is reflected in console output during CSV validation as well as in the CliApplication class itself. The feature fetches the version from a new ".version" file, unless it's operating in a PHPunit environment, in which case it outputs a generic 'unknown version' message.