Add delay to wait for Docker image and extend release workflows
This commit introduces a delay in the demo GitHub Action workflow to diminish the issue of the Docker image not being available immediately after release. Additionally, the workflows have been expanded on with more conditions for pushing to Docker and testing PHAR files. Lastly, some trivial changes have been made to output content for improved readability.