Remove delay block from demo workflow
This commit removes the unnecessary delay block which was waiting for the Docker image to build from the demo GitHub workflow. This change simplifies the workflow configuration and improves the execution time by eliminating an artificial 15-minute wait. Reference of the `jbzoo/csv-blueprint` tool in the main workflow was also updated to enhance stability.