Remove unnecessary conversion to float in aggregate methods
The methods stringsToFloat conversion in aggregate rules has been removed as it was unnecessary. The code for data processing has been refactored, removing the conversion of strings to float in the ComboMeanAbsDev, ComboMedianAbsDev and ComboQuartiles methods, thereby streamlining the code and potentially improving performance. Further, an exception has been added to the AbstractRule to handle non-existent method "validateRule".