Refine debugging and optimize benchmarks
Updated the debugging phrases in ValidatorCsv.php for better clarity in debug speed tracking. Benchmarks configuration in benchmarks.yml files were also adjusted for more efficient testing with renames for better identification, adjusted rules, as well as a reduction in the number of iterations and increased number of source rows. The Makefile was also updated to reflect these changes in benchmark file names.