Update README and test data values
The values of both test data in ReadmeTest.php and data presentation in README.md have been updated. The changes were based on v0.39 of the Docker - latest on March 30, 2024. These updates to the data values ensure that the performance parameters match the updated measurements from the tests, leading to more accurate results and readouts.