Add date interval validation rules
Updated the data validation code to include date interval checks. This added multiple new lines in JSON, PHP, YML and YAML file types to define these rules. In addition, a new file implementing the date interval validation logic was created, as well as a test file to ensure this feature works as expected.