Add 'is_isbn' validation rule (#142)
Introduced a new cell validation rule named `is_isbn` to validate if a
value is a valid International Standard Book Number (ISBN). The rule has
been added to the schema-examples and implemented in a new class file.
The README.md file and relevant test files have been updated to reflect
this change.