Add new sample and usage schemas and update ContainsNone rule
Two new schemas for sample and usage presets were added under the 'schema-examples' directory. These presets contain common validation rules for user data. In the ContainsNone rule, the error message was updated for better clarity by changing the phrase from containing "any of the following" to "the string". This makes it clear that a particular string is being contained, which improves readability and understandability of error messages.