Add ability to dump final schema
Added an option to dump the final schema of the CSV file for inspection after all includes and inheritance. Implementation involves modification in Commands classes to use this new option, and within the Schema class to convert the schema data to a string in YAML format. Documentation and relevant test files have been updated to reflect this change.