Implement `apply-all` option for global schema application without glob patterns (#150)
Renamed the `apply-global` command option to `apply-all` in CSV
validation, and adjusted all related usage within the documentation,
test, and action files. This change aims to improve clarity and
simplicity of command usage.

The 'apply-all' option now automates the application of a single schema
as global rules if no glob pattern is used. This update refines the
command's functionality and improves user experience. Changes have been
reflected in the documentation, test, and action files.