Refactor code for opcache preloading
This commit refactors the testing of CSV files and the configuration of the opcode cache (opcache). It deploys a PHP script that preloads all the classes available in the class map produced by Composer. Also, it introduces the creation of a `preload.php` file, which the updated Dockerfile then uses for opcache preloading, thereby enhancing the application's performance.