Upgrade PHP version and update dependencies
Updated the required PHP version to 8.2 for compatibility and enhanced performance characteristics. Updated Symfony packages and elevated giggsey/libphonenumber-for-php-lite library to the latest version. These changes ensure that the application is using the latest and safest versions of its dependencies, reducing potential security risks, and enchancing stability, performance, and interoperability.