Update error handling and CI configuration
The SonarCloud scan step in GitHub workflow is now allowed to continue on error. Updated error message in Utils.php for clarity and precision. Also, modified the Docker image source in action.yml to use the local Dockerfile. This change provides more flexibility and control over the docker environment for the actions.