Update benchmark setup in GitHub workflow and Makefile
The GitHub workflow and Makefile have been updated for a better benchmark setup. A separate step for multi-threaded benchmarking has been added and the method of generating random integers in the CSV file has been altered. The adjustments provide a clearer understanding of the process and better use of resources.