Adjust static calls in multiple rules (#162)
This commit removes the redundant artifact upload step from the GitHub
Actions workflow as it was unnecessary for the process. Also, the
self::compare() calls have been replaced with static::compare() in
AbstractAggregateRuleCombo, AbstractCellRuleCombo, and the pattern for
BIC number verification in IsBic rule has been updated for better
matching. Changes also include commenting on the NOSONAR issue where
applicable.